refactor: 컴포넌트 분리 (components/index.tsx → shared/ui/*)
