+++ src/shared/ui/index.ts
@@ -1,0 +1,27 @@
+export { Pagination } from './Pagination'
+export { Button } from './Button'
+export { Input } from './Input'
+export { Textarea } from './Textarea'
+export { Card, CardHeader, CardTitle, CardContent } from './Card'
+export {
+  Select,
+  SelectTrigger,
+  SelectValue,
+  SelectContent,
+  SelectItem,
+} from './Select'
+export {
+  Dialog,
+  DialogHeader,
+  DialogTrigger,
+  DialogContent,
+  DialogTitle,
+} from './Dialog'
+export {
+  Table,
+  TableHeader,
+  TableHead,
+  TableBody,
+  TableRow,
+  TableCell,
+} from './Table'
